MyAnimeArchive-server/myanimearchive/requirements.txt
Flask
celery[redis]
Flask-Cors==3.0.10
mysql-connector-python==8.0.28
python-decouple==3.5